requirements.txt

fastapi
uvicorn
pydantic
google-genai
aiohttp
python-multipart
